move filesystem to unstructured_data
