+++ pyproject.toml
@@ -65,11 +65,6 @@ facebook-business = "^16.0.2"
 [tool.poetry.group.salesforce.dependencies]
 simple-salesforce = "^1.12.4"
 
-[tool.poetry.group.filesystem.dependencies]
-google-api-python-client = "^2.86.0"
-google-auth-oauthlib = "^1.0.0"
-google-auth-httplib2 = "^0.1.0"
-
 
 [tool.poetry.group.unstructured_data.dependencies]
 langchain = "^0.0.219"
@@ -77,6 +72,7 @@ unstructured = "^0.7.10"
 openai = "^0.27.8"
 chromadb = "^0.3.26"
 tiktoken = "^0.4.0"
+
 
 [build-system]
 requires = ["poetry-core"]
