+++ sources/unstructured_data/__init__.py
@@ -5,8 +5,8 @@ import os
 from typing import Dict
 
 import dlt
-from dlt.extract.source import DltResource, TDataItem
 from dlt.common import logger
+from dlt.extract.source import DltResource, TDataItem
 
 from .helpers import (
     aprocess_file_to_structured,
