+++ sources/unstructured_data/filesystem/settings.py
@@ -5,7 +5,7 @@ MIME_TYPE_MAPPER = {".txt": "text/plain", ".pdf": "application/pdf"}
 ClIENT_SECRET_PATH = "client_secret.json"
 AUTHORIZED_USER_PATH = "token.json"
 STORAGE_FOLDER_PATH = "temp"
-FOLDER_IDS = ["1-yiloGjyl9g40VguIE1QnY5tcRPaF0Nm"]
+FOLDER_IDS = ["google_drive_folder_id",]
 
 # Local folder
 DATA_DIR = "test_data"
