+++ tests/unstructured_data/test_filesystem_source.py
@@ -1,11 +1,10 @@
 import tempfile
-from pathlib import Path
-from typing import List
+from typing import Sequence
 
 import dlt
 import pytest
 
-from sources.filesystem import google_drive, local_folder
+from sources.unstructured_data.filesystem import google_drive, local_folder
 
 from tests.utils import ALL_DESTINATIONS, assert_load_info
 
@@ -26,11 +25,6 @@ def run_pipeline(destination_name: str, resource):
 
 @pytest.mark.parametrize("destination_name", ALL_DESTINATIONS)
 class TestLoadFromLocalFolder:
-    @pytest.fixture
-    def data_dir(self) -> str:
-        current_dir = Path(__file__).parent.resolve()
-        return (current_dir / "test_data").as_posix()
-
     def test_load_info(self, destination_name: str, data_dir: str) -> None:
         # use extensions to filter files as 'extensions=(".txt", ".pdf", ...)'
         data_resource = local_folder(data_dir=data_dir)
@@ -65,29 +59,29 @@ class TestLoadFromLocalFolder:
 
 @pytest.mark.parametrize("destination_name", ALL_DESTINATIONS)
 class TestLoadFromGoogleDrive:
-    FOLDER_IDS: List[str] = ["1-yiloGjyl9g40VguIE1QnY5tcRPaF0Nm"]
-
     @pytest.mark.parametrize("download", [False, True])
-    def test_load_info(self, destination_name: str, download: bool) -> None:
+    def test_load_info(
+        self, destination_name: str, download: bool, gd_folders: Sequence[str]
+    ) -> None:
         with tempfile.TemporaryDirectory() as tmp_dir_path:
             # use extensions to filter files as 'extensions=(".txt", ".pdf", ...)'
             data_resource = google_drive(
                 download=download,
                 extensions=(".txt", ".pdf", ".jpg"),
                 storage_folder_path=tmp_dir_path,
-                folder_ids=self.FOLDER_IDS,
+                folder_ids=gd_folders,
             )
             _, load_info = run_pipeline(destination_name, data_resource)
             # make sure all data were loaded
             assert_load_info(load_info)
 
-    def test_tables(self, destination_name: str) -> None:
+    def test_tables(self, destination_name: str, gd_folders: Sequence[str]) -> None:
         with tempfile.TemporaryDirectory() as tmp_dir_path:
             # use extensions to filter files as 'extensions=(".txt", ".pdf", ...)'
             data_resource = google_drive(
                 download=True,
                 storage_folder_path=tmp_dir_path,
-                folder_ids=self.FOLDER_IDS,
+                folder_ids=gd_folders,
             )
             pipeline, _ = run_pipeline(destination_name, data_resource)
             # now let's inspect the generated schema. it should contain just
@@ -99,13 +93,15 @@ class TestLoadFromGoogleDrive:
             filepaths_table = tables[0]
             assert filepaths_table["name"] == "google_drive"
 
-    def test_google_drive_content(self, destination_name: str) -> None:
+    def test_google_drive_content(
+        self, destination_name: str, gd_folders: Sequence[str]
+    ) -> None:
         with tempfile.TemporaryDirectory() as tmp_dir_path:
             # use extensions to filter files as 'extensions=(".txt", ".pdf", ...)'
             data_resource = google_drive(
                 download=True,
                 storage_folder_path=tmp_dir_path,
-                folder_ids=self.FOLDER_IDS,
+                folder_ids=gd_folders,
             )
             pipeline, _ = run_pipeline(destination_name, data_resource)
             with pipeline.sql_client() as c:
